requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
rank-bm25>=0.2.2
json-repair>=0.8.0
requests>=2.31.0
